fix: move getHasher to TerserPlugin
